youtube_channel_transcription/requirements.txt
Flask
google-api-python-client
yt-dlp
openai-whisper
python-dotenv
gunicorn
